Article delete

+++ project-01-rest-jdbctemplate/src/main/java/com/example/project01restjdbctemplate/model/Article.java
@@ -22,11 +22,11 @@ public class Article {
 		this.body = body;
 	}
 
-	public long getId() {
+	public int getId() {
 		return id;
 	}
 
-	public void setId(long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
